리렌더링 방지 위한 컴포넌트 수정
