+++ src/app/components/login/CreateNewAccount.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useRouter } from "next/navigation";
 
 const CreateNewAccount = () => {
