+++ src/app/components/login/LoginSubmit.tsx
@@ -1,0 +1,58 @@
+"use client";
+
+import LoginForm from "./loginsubmit/LoginForm";
+import LoginButton from "./loginsubmit/LoginButton";
+import { useRouter } from "next/navigation";
+import { RootState } from "../../redux/reducers/index";
+import { useSelector, useDispatch } from "react-redux";
+import { logIn, logInInput } from "../../redux/actions/userAction";
+
+export interface logInFormData {
+  phoneNumberOrEmail: string;
+  password: string;
+}
+
+const LogInSubmit = () => {
+  const { login } = useSelector((state: RootState) => ({
+    login: state.user.login,
+  }));
+  // console.log("login", login);
+  const dispatch = useDispatch();
+
+  // 폼 입력 상태 업데이트
+  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = event.target;
+    // console.log(login);
+    dispatch(logInInput({ ...login, [name]: value }));
+  };
+
+  // router 객체
+  const router = useRouter();
+
+  // 폼 제출 핸들러
+  const onSubmit = (event?: React.FormEvent<HTMLFormElement>) => {
+    if (event) event.preventDefault();
+
+    // console.log("Submitted data:", login);
+
+    dispatch(logIn(login));
+    dispatch(
+      logInInput({
+        phoneNumberOrEmail: "",
+        password: "",
+      })
+    );
+
+    // 폼 제출 투두리스트 페이지로 이동
+    router.push("/todolist");
+  };
+
+  return (
+    <>
+      <LoginForm formData={login} onChange={onChange} />
+      <LoginButton onSubmit={onSubmit} />
+    </>
+  );
+};
+
+export default LogInSubmit;
