+++ src/app/components/todolist/ToDoTitle.tsx
@@ -2,7 +2,7 @@ const ToDoTitle = () => {
   return (
     <>
       {/* To-Do List Title */}
-      <div className="ml-[24px] mt-[32px] font-pretendard font-semibold text-[24px] text-[#323233]">
+      <div className="mx-[24px] mt-[32px] font-pretendard font-semibold text-[24px] text-[#323233]">
         To-Do List
       </div>
     </>
