+++ src/app/components/todolist/todoitems/ToDoListItem.tsx
@@ -21,8 +21,12 @@ const TodoListItem = ({ todo, handler }: TodoListItemProps) => {
 
   // 수정 버튼 클릭
   const handleUpdateClick = () => {
+    console.log(todo.isEditing);
+    console.log(todo);
+
     todo.newTitle === ""
-      ? dispatch(setNewTitle(todo.id, todo.title))
+      ? dispatch(setNewTitle(todo.id, todo.title)) &&
+        dispatch(toggleEditing(todo.id, todo.isEditing))
       : dispatch(toggleEditing(todo.id, todo.isEditing));
 
     todo.isEditing &&
@@ -68,7 +72,7 @@ const TodoListItem = ({ todo, handler }: TodoListItemProps) => {
       )}
       {/* todo update button */}
       <button className="w-[24px] h-[24px]" onClick={handleUpdateClick}>
-        <Image src={deleteButton} alt={deleteButton} width={24} height={24} />
+        <Image src={deleteButton} alt="deleteButton" width={24} height={24} />
       </button>
       {/* todo delete button */}
       <button
@@ -77,7 +81,7 @@ const TodoListItem = ({ todo, handler }: TodoListItemProps) => {
           handler.onDelete(todo.id);
         }}
       >
-        <Image src={deleteButton} alt={deleteButton} width={24} height={24} />
+        <Image src={deleteButton} alt="deleteButton" width={24} height={24} />
       </button>
     </li>
   );
